Fix new error w/ HA 2024.8
Move YAML config processing to config.py.
A couple other minor fixes/improvements.
