+++ custom_components/__init__.py
@@ -1,0 +1,2 @@
+"""Composite Device Tracker."""
+# Exists to satisfy mypy.
